refactor any2mochi parsing
